nomnom-backend/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
Flask-Bcrypt
Flask-JWT-Extended
gunicorn
pandas
openpyxl
python-dotenv
psycopg2-binary
scikit-learn
scikit-surprise